monomial manager refactoring
