+++ triples/utils/basis_generator.py
@@ -1,294 +1,6 @@
-from typing import Union, Optional, Dict, List, Tuple, Callable
+from warnings import warn
 
-import numpy as np
-import sympy as sp
-from sympy.combinatorics import Permutation, PermutationGroup, CyclicGroup, SymmetricGroup
+from .monomials import *
 
-def generate_partitions(nvars: int, degree: int, equal: bool = False) -> List[Tuple[int, ...]]:
-    """
-    Generate all tuples (a0,a1,...,an) such that n = nvars and sum(ai) <= degree.
-    If equal is True, then it requires sum(ai) == degree.
-    """
-    if nvars == 0:
-        return [()] if degree == 0 or (not equal) else []
-    def generate_tuples(current_tuple: Tuple[int, ...], current_sum: int, remaining_vars: int) -> List[Tuple[int, ...]]:
-        if remaining_vars == 0:
-            return [current_tuple]
-        else:
-            tuples = []
-            for i in range(degree - current_sum, -1, -1):
-                tuples.extend(generate_tuples(current_tuple + (i,), current_sum + i, remaining_vars - 1))
-            return tuples
-
-    monoms = generate_tuples((), 0, nvars) if not equal else generate_tuples((), 0, nvars - 1)
-    if equal:
-        monoms = [m + (degree - sum(m),) for m in monoms]
-    return monoms
-    # inv_monoms = list(filter(self.is_standard_monom, inv_monoms))
-    # dict_monoms = {t: i for i, t in enumerate(inv_monoms)}
-    # return dict_monoms, inv_monoms
-
-
-class MonomialManager():
-    """
-    Class to compute polynomial monomials given the symmetry of variables and homogeneity.
-    """
-    def __init__(self, nvars: int, perm_group = None, is_homogeneous: bool = True) -> None:
-        self.nvars = nvars
-        self._is_homogeneous = bool(is_homogeneous)
-        if isinstance(perm_group, MonomialManager):
-            perm_group = perm_group.perm_group
-        self._perm_group = perm_group if perm_group is not None else PermutationGroup(Permutation(list(range(nvars))))
-        if self._perm_group.degree != nvars:
-            raise ValueError(f"The degree of the permutation group ({self._perm_group.degree}) does not match the number of variables ({nvars}).")
-
-        self._monoms = {}
-
-    def __str__(self) -> str:
-        return "MonomialManager(nvars=%d, perm_group=%s, is_homogeneous=%s)" % (
-            self.nvars, str(self._perm_group).replace('\n', '').replace('  ',''), self._is_homogeneous)
-
-    def __repr__(self) -> str:
-        return self.__str__()
-
-    def copy(self) -> 'MonomialManager':
-        return MonomialManager(self.nvars, perm_group=self._perm_group, is_homogeneous=self._is_homogeneous)
-
-    @property
-    def is_homogeneous(self) -> bool:
-        return self._is_homogeneous
-
-    @property
-    def perm_group(self) -> PermutationGroup:
-        return self._perm_group
-
-    @property
-    def is_trivial(self) -> bool:
-        return self._perm_group.is_trivial
-
-    @classmethod
-    def from_perm_group(cls, perm_group: PermutationGroup, is_homogeneous: bool = True) -> 'MonomialManager':
-        """
-        Create a `MonomialManager` object from a permutation group. The `nvars` argument is inferred from 
-        the degree of the group and is therefore not needed.
-        """
-        return MonomialManager(perm_group.degree, perm_group=perm_group, is_homogeneous=is_homogeneous)
-
-    def _register_monoms(self, degree: int) -> Tuple[Dict[Tuple[int, ...], int], List[Tuple[int, ...]]]:
-        """
-        Register dict_monoms and inv_monoms for degree if not computed.
-        """
-        x = self._monoms.get(degree, None)
-        if x is None:
-            candidates = generate_partitions(self.nvars, degree, equal = self._is_homogeneous)
-            monoms = list(filter(self.is_standard_monom, candidates))
-            dict_monoms = {t: i for i, t in enumerate(monoms)}
-            x = (dict_monoms, monoms)
-            self._monoms[degree] = x
-        return x
-
-    def dict_monoms(self, degree: int) -> Dict[Tuple[int, ...], int]:
-        """
-        Return the dictionary of monomials of given degree with indices as values.
-        """
-        return self._register_monoms(degree)[0]
-
-    def inv_monoms(self, degree: int) -> List[Tuple[int, ...]]:
-        """
-        Return the list of monomials of given degree. Monomials can be accessed by indices.
-        """
-        return self._register_monoms(degree)[1]
-
-    def index(self, monom: Tuple[int, ...]) -> Optional[int]:
-        """
-        Return the index of the monom in the vector representation of the polynomial.
-        Note that it is NOT the index of a permutation group.
-        """
-        degree = sum(monom)
-        return self.dict_monoms(degree).get(self.standard_monom(monom))
-
-    def length(self, degree: int) -> int:
-        return len(self.inv_monoms(degree))
-
-    def __contains__(self, monom: Tuple[int, ...]) -> bool:
-        """
-        Return True if a monom is contained in the reduction rule.
-        """
-        return monom in self.dict_monoms(sum(monom))
-
-    def base(self) -> 'MonomialManager':
-        """
-        Return a copy with the permutation group set to the trivial group.
-        """
-        obj = self.copy()
-        obj._perm_group = PermutationGroup(Permutation(list(range(self.nvars))))
-        return obj
-
-    def order(self) -> int:
-        """
-        Return the order of the permutation group.
-        """
-        return self._perm_group.order()
-
-    def permute(self, monom: Tuple[int, ...]) -> List[Tuple[int, ...]]:
-        """
-        Permute the monom according to the permutation group.
-        """
-        return [tuple(perm(monom)) for perm in self._perm_group.elements]
-
-    def standard_monom(self, monom: Tuple[int, ...]) -> Tuple[int, ...]:
-        """
-        For monomials in cyclic groups, only one representative is needed.
-        For example, a^3b^2c, b^3c^2a, and c^3a^2b are equivalent wrt. CyclicGroup(3),
-        and the standard monom (the representative) is (3, 2, 1), standing for a^3b^2c.
-        The function returns the standard monom for the input monom. 
-        This is chosen to be the lexicographically largest monom among all permutations of the input monom.
-        """
-        return max(self.permute(monom))
-
-    def is_standard_monom(self, monom: Tuple[int, ...]) -> bool:
-        """
-        For monoms in cyclic groups, only one representative is needed.
-        For example, a^3b^2c, b^3c^2a, and c^3a^2b are equivalent in MonomialCyclic,
-        and the standard monom (the representative) is a^3b^2c. The function
-        returns True if the input monom is the standard monom.
-        """
-        return monom == self.standard_monom(monom)
-
-    def permute_vec(self, vec: sp.Matrix, degree: int) -> sp.Matrix:
-        """
-        Permute the vector representation of a polynomial according to the permutation group.
-        It returns a matrix with all permutations of the monoms.
-
-        Note that the vector representation is on the base of MonomialManager class, which does
-        not reduce any monomials by symmetry.
-        """
-        base = self.base()
-        dict_monoms = base.dict_monoms(degree)
-        inv_monoms = base.inv_monoms(degree)
-        f = lambda i: dict_monoms[i]
-        # get all permutations of all monoms first
-        all_vecs = [[0] * vec.shape[0] for _ in range(self.order())]
-        for v, m in zip(vec, inv_monoms):
-            for ind, j in enumerate(map(f, self.permute(m))):
-                # write the value to the corresponding column
-                all_vecs[ind][j] = v
-        return sp.Matrix(all_vecs).T
- 
-
-    # def _standard_monom(self, monom: Tuple[int, ...]) -> Tuple[int, ...]:
-    #     warn("_standard_monom is deprecated. Use standard_monom instead.", DeprecationWarning, stacklevel=2)
-    #     return self.standard_monom(monom)
-
-    def _assert_equal_nvars(self, t: Tuple[int, ...]) -> bool:
-        if len(t) != self.nvars:
-            raise ValueError("Number of variables does not match. Expected %d but received %d." % (self.nvars, len(t)))
-            return False
-        return True
-
-    def _arraylize_list(self, poly: sp.Poly, expand_cyc: bool = False) -> List[int]:
-        self._assert_equal_nvars(poly.gens)
-        degree = poly.total_degree()
-        dict_monoms = self.dict_monoms(degree)
-        array = [0] * len(dict_monoms)
-        if not expand_cyc:
-            for monom, coeff in poly.terms():
-                v = dict_monoms.get(monom)
-                if v is not None:
-                    array[v] = coeff
-        else:
-            for monom, coeff in poly.terms():
-                for monom2 in self.permute(monom):
-                    v = dict_monoms.get(monom2)
-                    if v is not None:
-                        array[v] += coeff
-        return array
-
-    def arraylize(self, poly: sp.Poly, expand_cyc: bool = False) -> np.ndarray:
-        """
-        Return the vector representation of the polynomial in numpy array.
-        """
-        return np.array(self._arraylize_list(poly, expand_cyc = expand_cyc)).astype(np.float64)
-
-    def arraylize_sp(self, poly: sp.Poly, expand_cyc: bool = False) -> sp.Matrix:
-        """
-        Return the vector representation of the polynomial in sympy matrix (column vector).
-        """
-        return sp.Matrix(self._arraylize_list(poly, expand_cyc = expand_cyc))
-
-    def invarraylize(self, array: Union[List, np.ndarray, sp.Matrix], gens: List[sp.Symbol], degree: int) -> sp.Poly:
-        """
-        Reverse the arraylize function to get the polynomial from its vector representation.
-        """
-        self._assert_equal_nvars(gens)
-        inv_monoms = self.inv_monoms(degree)
-        terms_dict = {}
-        for coeff, monom in zip(array, inv_monoms):
-            for monom2 in self.permute(monom):
-                terms_dict[monom2] = coeff
-        return sp.Poly(terms_dict, gens)
-
-    def cyclic_sum(self, expr: sp.Expr, gens: List[sp.Symbol]) -> sp.Expr:
-        """
-        Sum up a given expression according to the permutation group.
-
-        This function should be hijacked by the CyclicExpr module.
-        """
-        s = []
-        expr0 = expr
-        for perm in self._perm_group.elements:
-            perm_gens = perm(gens)
-            expr = expr0.xreplace(dict(zip(gens, perm_gens)))
-            s.append(expr)
-        return sp.Add(*s)
-
-
-def _parse_options(nvars, **options) -> MonomialManager:
-    if nvars < 0:
-        raise ValueError("Number of variables should be non-negative.")
-
-    hom = options.pop('hom', True)
-    cyc = options.pop('cyc', False)
-    sym = options.pop('sym', False)
-    symmetry = options.pop('symmetry', None)
-
-    if len(options) > 0:
-        raise KeyError(f"Unknown options: {','.join(options.keys())}."
-                       ' Only "hom", "cyc", "sym", and "symmetry" are allowed.')
-
-    if symmetry is not None and (cyc or sym):
-        raise ValueError("Cannot specify both symmetry and cyc or sym.")  
-    elif cyc:
-        if sym:
-            raise ValueError("Cannot specify both cyc and sym.")
-        symmetry = CyclicGroup(nvars)
-    elif sym:
-        symmetry = SymmetricGroup(nvars)
-
-    if isinstance(symmetry, MonomialManager):
-        return symmetry
-    if symmetry is None:
-        return MonomialManager(nvars, is_homogeneous = hom)
-    if isinstance(symmetry, PermutationGroup):
-        return MonomialManager.from_perm_group(symmetry, is_homogeneous = hom)
-    
-    raise ValueError(f"Invalid symmetry type {type(symmetry)}. Expected MonomialManager or PermutationGroup.")
-
-
-def arraylize(poly: sp.Poly, expand_cyc: bool = False, **options) -> np.ndarray:
-    option = _parse_options(len(poly.gens), **options)
-    return option.arraylize(poly, expand_cyc = expand_cyc)
-
-def arraylize_sp(poly: sp.Poly, expand_cyc: bool = False, **options) -> sp.Matrix:
-    option = _parse_options(len(poly.gens), **options)
-    return option.arraylize_sp(poly, expand_cyc = expand_cyc)
-
-def invarraylize(array: Union[List, np.ndarray, sp.Matrix], gens: List[sp.Symbol], degree: int, **options) -> sp.Poly:
-    option = _parse_options(len(gens), **options)
-    return option.invarraylize(array, gens, degree)
-
-def generate_expr(nvars: int, degree: int, **options) -> Tuple[Dict[Tuple[int, ...], int], List[Tuple[int, ...]]]:
-    option = _parse_options(nvars, **options)
-    return option._register_monoms(degree)
-
-
+warn("Importing basis_generator from triples.utils is deprecated. Please import from triples.utils.monomials instead.", DeprecationWarning,
+     stacklevel=2)
